requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.7.0=pypi_0
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.13=pypi_0
aiosignal=1.4.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.9.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
audioread=3.0.1=pypi_0
av=15.0.0=pypi_0
blake3=1.0.5=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=6.2.0=pypi_0
cbor2=5.7.0=pypi_0
certifi=2025.7.9=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.1=pypi_0
cloudpickle=3.1.1=pypi_0
compressed-tensors=0.10.2=pypi_0
contourpy=1.3.2=pypi_0
cupy-cuda12x=13.6.0=pypi_0
cycler=0.12.1=pypi_0
datasets=3.6.0=pypi_0
decorator=5.2.1=pypi_0
depyf=0.19.0=pypi_0
dill=0.3.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
dnspython=2.8.0=pypi_0
docstring-parser=0.16=pypi_0
einops=0.8.1=pypi_0
email-validator=2.3.0=pypi_0
exceptiongroup=1.3.0=pypi_0
expat=2.7.1=h6a678d5_0
fastapi=0.116.0=pypi_0
fastapi-cli=0.0.10=pypi_0
fastapi-cloud-cli=0.1.5=pypi_0
fastrlock=0.8.3=pypi_0
ffmpy=0.6.0=pypi_0
filelock=3.18.0=pypi_0
fire=0.7.0=pypi_0
flask=3.1.2=pypi_0
fonttools=4.58.5=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.3.0=pypi_0
gguf=0.17.1=pypi_0
gradio=5.31.0=pypi_0
gradio-client=1.10.1=pypi_0
groovy=0.1.2=pypi_0
h11=0.16.0=pypi_0
hf-transfer=0.1.9=pypi_0
hf-xet=1.1.5=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.34.4=pypi_0
idna=3.10=pypi_0
imagehash=4.3.2=pypi_0
interegular=0.3.3=pypi_0
itsdangerous=2.2.0=pypi_0
jieba=0.42.1=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
joblib=1.5.1=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
kiwisolver=1.4.8=pypi_0
lark=1.2.2=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
librosa=0.11.0=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
llamafactory=0.9.4.dev0=pypi_0
llguidance=0.7.30=pypi_0
llvmlite=0.44.0=pypi_0
lm-format-enforcer=0.10.12=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
mdurl=0.1.2=pypi_0
mistral-common=1.8.4=pypi_0
modelscope=1.28.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.1=pypi_0
msgspec=0.19.0=pypi_0
multidict=6.6.3=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
ninja=1.13.0=pypi_0
nltk=3.9.1=pypi_0
numba=0.61.2=pypi_0
numpy=2.2.6=pypi_0
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.106.1=pypi_0
openai-harmony=0.0.4=pypi_0
opencv-python=4.12.0.88=pypi_0
opencv-python-headless=4.12.0.88=pypi_0
openssl=3.0.16=h5eee18b_0
orjson=3.10.18=pypi_0
outlines-core=0.2.10=pypi_0
packaging=25.0=pypi_0
pandas=2.3.1=pypi_0
partial-json-parser=0.2.1.1.post6=pypi_0
peft=0.15.2=pypi_0
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pypi_0
polars=1.33.1=pypi_0
pooch=1.8.2=pypi_0
prometheus-client=0.22.1=pypi_0
prometheus-fastapi-instrumentator=7.1.0=pypi_0
propcache=0.3.2=pypi_0
protobuf=6.31.1=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
py-cpuinfo=9.0.0=pypi_0
pyarrow=20.0.0=pypi_0
pybase64=1.4.2=pypi_0
pycountry=24.6.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pydantic-extra-types=2.10.5=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.2.3=pypi_0
pytesseract=0.3.13=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.1=pypi_0
python-json-logger=3.3.0=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.2=pypi_0
pywavelets=1.8.0=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=27.0.2=pypi_0
qwen-vl-utils=0.0.11=pypi_0
ray=2.49.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.4=pypi_0
rich=14.0.0=pypi_0
rich-toolkit=0.15.1=pypi_0
rignore=0.6.4=pypi_0
rouge-chinese=1.0.3=pypi_0
rpds-py=0.27.1=pypi_0
ruff=0.12.2=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.7.0=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
sentence-transformers=5.1.2=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.37.0=pypi_0
setproctitle=1.3.7=pypi_0
setuptools=78.1.1=py310h06a4308_0
shellingham=1.5.4=pypi_0
shtab=1.7.2=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soundfile=0.13.1=pypi_0
soxr=0.5.0.post1=pypi_0
sqlite=3.45.3=h5eee18b_0
sse-starlette=2.4.1=pypi_0
starlette=0.46.2=pypi_0
sympy=1.14.0=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
timm=1.0.19=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.22.0=pypi_0
tomlkit=0.13.3=pypi_0
torch=2.7.1=pypi_0
torchaudio=2.7.1=pypi_0
torchvision=0.22.1=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.56.1=pypi_0
triton=3.3.1=pypi_0
trl=0.9.6=pypi_0
typer=0.16.0=pypi_0
typing-extensions=4.14.1=pypi_0
typing-inspection=0.4.1=pypi_0
tyro=0.8.14=pypi_0
tzdata=2025.2=pypi_0
ultralytics=8.3.197=pypi_0
ultralytics-thop=2.0.17=pypi_0
urllib3=2.5.0=pypi_0
uvicorn=0.35.0=pypi_0
uvloop=0.21.0=pypi_0
vllm=0.10.1.1=pypi_0
watchfiles=1.1.0=pypi_0
websockets=15.0.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xformers=0.0.31=pypi_0
xgrammar=0.1.21=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.5.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
